requirements.txt
fastapi
uvicorn
eodash_catalog
requests
GitPython
pydantic-settings
python-multipart
PyYAML
pytest
